+++ src/main/java/org/opentripplanner/standalone/config/RouterConfig.java
@@ -27,7 +27,6 @@ import org.slf4j.LoggerFactory;
  */
 public class RouterConfig implements Serializable {
 
-  private static final Duration DEFAULT_STREET_ROUTING_TIMEOUT = Duration.ofSeconds(5);
   private static final Logger LOG = LoggerFactory.getLogger(RouterConfig.class);
 
   public static final RouterConfig DEFAULT = new RouterConfig(
@@ -43,7 +42,6 @@ public class RouterConfig implements Serializable {
   private final String configVersion;
   private final String requestLogFile;
   private final TransmodelAPIConfig transmodelApi;
-  private final Duration streetRoutingTimeout;
   private final RouteRequest routingRequestDefaults;
   private final TransitRoutingConfig transitConfig;
   private final UpdatersParameters updatersParameters;
@@ -105,13 +103,18 @@ number of transit vehicles used in that itinerary.
           .summary("Configuration for the Transmodel GraphQL API.")
           .asObject()
       );
-    this.streetRoutingTimeout = parseStreetRoutingTimeout(root);
     this.routingRequestDefaults =
       RouteRequestConfig.mapDefaultRouteRequest(root, "routingDefaults");
     this.transitConfig = new TransitRoutingConfig("transit", root, routingRequestDefaults);
     this.updatersParameters = new UpdatersConfig(root);
     this.vectorTileLayers = VectorTileConfig.mapVectorTilesParameters(root, "vectorTileLayers");
     this.flexConfig = new FlexConfig(root, "flex");
+
+    this.routingRequestDefaults.withPreferences(p ->
+        p.withStreet(s ->
+          s.withRoutingTimeout(parseStreetRoutingTimeout(root, s.original().routingTimeout()))
+        )
+      );
 
     if (logUnusedParams && LOG.isWarnEnabled()) {
       root.logAllUnusedParameters(LOG::warn);
@@ -137,15 +140,6 @@ number of transit vehicles used in that itinerary.
 
   public String requestLogFile() {
     return requestLogFile;
-  }
-
-  /**
-   * The preferred way to limit the search is to limit the distance for each street mode(WALK, BIKE,
-   * CAR). So the default timeout for a street search is set quite high. This is used to abort the
-   * search if the max distance is not reached within the timeout.
-   */
-  public Duration streetRoutingTimeout() {
-    return streetRoutingTimeout;
   }
 
   public TransmodelAPIConfig transmodelApi() {
@@ -198,7 +192,7 @@ number of transit vehicles used in that itinerary.
    *
    * @since 2.2 - The support for the old format can be removed in version > 2.2.
    */
-  static Duration parseStreetRoutingTimeout(NodeAdapter adapter) {
+  static Duration parseStreetRoutingTimeout(NodeAdapter adapter, Duration defaultValue) {
     return adapter
       .of("streetRoutingTimeout")
       .since(NA)
@@ -219,6 +213,6 @@ search-window.
 The search aborts after this duration and any paths found are returned to the client.
 """
       )
-      .asDuration(DEFAULT_STREET_ROUTING_TIMEOUT);
+      .asDuration(defaultValue);
   }
 }