+++ src/main/java/org/opentripplanner/standalone/configure/ConstructApplicationModule.java
@@ -34,7 +34,6 @@ public class ConstructApplicationModule {
     return DefaultServerRequestContext.create(
       routerConfig.transitTuningConfig(),
       routerConfig.routingRequestDefaults(),
-      routerConfig.streetRoutingTimeout(),
       raptorConfig,
       graph,
       transitService,