+++ src/test/java/org/opentripplanner/transit/speed_test/SpeedTest.java
@@ -105,7 +105,6 @@ public class SpeedTest {
       DefaultServerRequestContext.create(
         config.transitRoutingParams,
         config.request,
-        null,
         new RaptorConfig<>(config.transitRoutingParams),
         graph,
         new DefaultTransitService(transitModel),